Expand aspects of a recolorize object for other functions

Expand aspects of a recolorize object for other functions

Usage

expand_recolorize(
  recolorize_obj,
  original_img = FALSE,
  recolored_img = FALSE,
  sizes = FALSE
)

Arguments

recolorize_obj

A recolorize object.

original_img

Logical. Return original image as numeric array?

recolored_img

Logical. Return recolored image as numeric array?

sizes

Logical. Return cluster sizes (as number of pixels)?

Value

A recolorize object with the indicated additional elements, as well as the original elements.